backend/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
python-dotenv==1.0.1
pydantic==2.9.0
requests==2.31.0
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.0
joblib==1.4.0
